Main contributor: Lars

Summary of additions:
- The wildfire model BLAZE (BLAZe induced land-atmosphere flux Estimator, 
  including the burned area simulator SIMFIRE) computing burned area 
  on a daily basis and computing the fluxes between live- and
  litter-pools and the atmosphere for C and N. 
- The Global WeatherGENerator (GWGEN, Sommer et al.) producing correlated 
  daily values from monthly input of:
  - mean temperature
  - daily temperature range
  - precipitation
  - solar downward short-wave radiation
  - 10m wind (new)
  - relative humidity (new)

Changes to input/added inputs:
* New meteorological variables relative humidity (or specific humidity 
  and pressure) and horizontal wind.
* Simfire input.
* Benchmark crop_global postprocess.sh specifies path to observational 
  fire data.

Changes to Output:
* 3 additional output files can be written
  - monthly burned area
  - annual burned area
  - simfireanalysis with annual data: simfire_biome, max. Nesterov-Index, 
    human population density
* Benchmarks:
  Crop_global:
  - comparison of average annual C-emissions by fire (1997-2013, map) to GFED 4.0.
  - comparison of average annual C-emissions by GFED-region (1997-2013, table) to GFED 4.0.
  Global (pnv):
  - A new map depicting the strength of dominance (in LAI) of the
    dominant pft has been added.  

Benchmark results:
* An increase in the agreement with the sought-after biome distribution from
  Haxeltine and Prentice (1996) was observed after the introduction of BLAZE
  as compared to the parametrisation (r8185).
* Crop_global:
  In general, the global fire-emissions have been reduced to about 1.4 Pg
  C/yr. This is a slight underestimation against e.g. GFED 4 which is
  due to the non-representation of peat-fires and an underrepresentation
  of pasture burns in the tropics, specifically in Africa.
  Overall the global patterns are reproduced.
* Panarctic:
  A change in active layer depth (larger area with parmafrost) that is mainly
  due to the change in wheather generator was observed. A similar change in vegetation
  carbon was also observed, with lower values throughout the artic and sub-arctic area,
  which then resulted in overall lower carbon pools.
* Wetlands:
  A similar change in response, as in panarctic, could be seen in the global methane 
  emissions. The new results (~175 TgC CH4) are more in line with the observations
  (~150 TgC CH4).

Speed/performance:
* The use of the weathergenerator GWGEN adds an approximate amount of 19% to a run 
  using PNV only (global benchmark), and about 10% run-time to a run using crops 
  (crop_global benchmark), respectively.

Issues when updating:
* A deviation by about 1% VegC (crop_global and global PNV) occurs when comparing 
  to trunk r7818 with this trunk-revision (ver 4.1) with BLAZE and GWGEN switched off. 
  The changes responsible have been applied to vegdynam.cpp and are mathematically 
  consistent with the previous version. They have been approved on developers' 
  meeting 2018-09-27. 
* This branch has taken part in the global retuning for release 4.1 and therefore 
  comparison with older versions of trunk should be viewed with caution.
  For a comprehensive review of diffences in  benchmarks results between trunk v4.0.1 
  vs the model after addition of 4.1 features and retuning of parameters, see the
  commit message of the shortly forthcoming reintegration to trunk of branch 
  nitrogen_transformations.
* Documentation for the blaze feature branch in guess_reference will be added
  in a later commit to trunk (then also documentation for Arctic & Wetlands feature
  will be added).

Branch history:
This branch was created from trunk r5985. 
Code from an older branch was added @5992.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/trunk@8216 70805675-57b6-444b-bb00-18ecffcb4be8

+++ modules/growth.cpp
@@ -1229,7 +1229,7 @@ void growth(Stand& stand, Patch& patch) {
 				// BLARP! excess allocation to roots now also included (assumes leaf longevity = root longevity)
 
 				cmass_excess = max((double)indiv.aphen_raingreen /
-					(indiv.pft.leaflong * date.year_length()) * (indiv.cmass_leaf + indiv.cmass_root) -
+					(indiv.pft.leaflong * (double)date.year_length()) * (indiv.cmass_leaf + indiv.cmass_root) -
 					indiv.cmass_leaf - indiv.cmass_root, 0.0);
 
 				if (cmass_excess > bminc) cmass_excess = bminc;
